Now curl libs not required.
