+++ BUILD.bazel
@@ -1,6 +1,28 @@
 cc_library(
     name = "dd_opentracing_cpp",
-    srcs = glob(["src/*.cpp", "src/*.h"]) + [
+    srcs = [
+        "src/clock.h",
+        "src/noopspan.cpp",
+        "src/noopspan.h",
+        "src/opentracing_external.cpp",
+        "src/propagation.cpp",
+        "src/propagation.h",
+        "src/publisher.cpp",
+        "src/publisher.h",
+        "src/sample.cpp",
+        "src/sample.h",
+        "src/span_buffer.cpp",
+        "src/span_buffer.h",
+        "src/span.cpp",
+        "src/span.h",
+        "src/tracer.cpp",
+        "src/tracer.h",
+        "src/transport.cpp",
+        "src/transport.h",
+        "src/version_check.cpp",
+        "src/version_check.h",
+        "src/writer.cpp",
+        "src/writer.h",
         ":version_number.h",
     ],
     hdrs = [
