+++ src/opentracing_agent.cpp
@@ -1,4 +1,5 @@
 #include <datadog/opentracing.h>
+#include "agent_writer.h"
 #include "tracer.h"
 
 namespace ot = opentracing;
@@ -7,16 +8,10 @@ namespace datadog {
 namespace opentracing {
 
 std::shared_ptr<ot::Tracer> makeTracer(const TracerOptions &options) {
-  return std::shared_ptr<ot::Tracer>{new Tracer{options}};
-}
-
-std::tuple<std::shared_ptr<ot::Tracer>, std::shared_ptr<TracePublisher>> makeTracerAndPublisher(
-    const TracerOptions &options) {
-  auto xwriter = std::make_shared<ExternalWriter>();
-  auto publisher = xwriter->publisher();
-  std::shared_ptr<Writer> writer = xwriter;
-  return std::tuple<std::shared_ptr<ot::Tracer>, std::shared_ptr<TracePublisher>>{
-      std::shared_ptr<ot::Tracer>{new Tracer{options, writer}}, publisher};
+  auto writer = std::shared_ptr<Writer>{
+      new AgentWriter(options.agent_host, options.agent_port,
+                      std::chrono::milliseconds(llabs(options.write_period_ms)))};
+  return std::shared_ptr<ot::Tracer>{new Tracer{options, writer}};
 }
 
 }  // namespace opentracing
