+++ src/tracer.cpp
@@ -16,13 +16,6 @@ uint64_t getId() {
   return distribution(source);
 }
 
-Tracer::Tracer(TracerOptions options)
-    : Tracer(options,
-             std::shared_ptr<SpanBuffer>{new WritingSpanBuffer{std::make_shared<AgentWriter>(
-                 options.agent_host, options.agent_port,
-                 std::chrono::milliseconds(llabs(options.write_period_ms)))}},
-             getRealTime, getId, ConstantRateSampler(options.sample_rate)) {}
-
 Tracer::Tracer(TracerOptions options, std::shared_ptr<SpanBuffer> buffer, TimeProvider get_time,
                IdProvider get_id, SampleProvider sampler)
     : opts_(options),
@@ -32,7 +25,10 @@ Tracer::Tracer(TracerOptions options, std::shared_ptr<SpanBuffer> buffer, TimePr
       sampler_(sampler) {}
 
 Tracer::Tracer(TracerOptions options, std::shared_ptr<Writer> &writer)
-    : opts_(options), get_time_(getRealTime), get_id_(getId), sampler_(KeepAllSampler()) {
+    : opts_(options),
+      get_time_(getRealTime),
+      get_id_(getId),
+      sampler_(ConstantRateSampler(options.sample_rate)) {
   buffer_ = std::shared_ptr<SpanBuffer>{new WritingSpanBuffer{writer}};
 }
