+++ test/CMakeLists.txt
@@ -14,4 +14,4 @@ _datadog_test(span_test span_test.cpp)
 _datadog_test(tracer_factory_test tracer_factory_test.cpp)
 _datadog_test(tracer_test tracer_test.cpp)
 _datadog_test(sample_test sample_test.cpp)
-_datadog_test(writer_test writer_test.cpp)
+_datadog_test(agent_writer_test agent_writer_test.cpp)
